requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
fastapi
uvicorn
pillow
pytz
gdal==3.10.2
django-leaflet==0.31.0
jdatetime
geojson-tiles==0.1.2
geojson==3.2.0
persiantools==5.2.0
reportlab
bidi-python
arabic-reshaper
django-geojson
geopy
celery
redis
django-celery-beat